requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-grappelli==3.0.3
django-jet-reboot==1.3.2
django-js-asset==2.0.0
idna==3.4
Pillow==9.2.0
psycopg2-binary==2.9.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
uWSGI==2.0.20
